Refactor Binance exchange helpers into dedicated modules
